requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiomysql==0.2.0
aiosignal==1.4.0
alembic==1.17.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
jmespath==1.0.1
loguru==0.7.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
pydantic==2.11.10
pydantic_core==2.33.2
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==1.17.3
yarl==1.22.0
